add basic validation to dish post request

+++ database/migrations/2025_04_14_094320_create_dishes_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name', 100);
             $table->text('description')->nullable();
-            $table->float('price');
+            $table->double('price', 8, 2);
             $table->string('image')->nullable();
 
             $table->timestamps();
